Switched from using a union for my temp fcf
I also made temp fcf private.

The temporary Frame Control Field is needed, because there are many options packed into two bytes.
I could, and might modify the frameControlField struct into a class with getters and setters.
If I did that, then I could operate on the data itself.
Since I'm currently using a struct, the data needs to be packed into those two bytes.

I was using a union for the temp fcf, but now I'm using pointers and references.
This isn't needed, but is good practice for when I use something similar with the address stuff.